Extract comment registrar
