+++ src/Service/FileProcessor.php
@@ -4,18 +4,13 @@ declare(strict_types=1);
 
 namespace Aeliot\TodoRegistrar\Service;
 
-use Aeliot\TodoRegistrar\Service\Comment\Detector as CommentDetector;
-use Aeliot\TodoRegistrar\Service\Comment\Extractor as CommentExtractor;
 use Aeliot\TodoRegistrar\Service\File\Saver;
 use Aeliot\TodoRegistrar\Service\File\Tokenizer;
-use Aeliot\TodoRegistrar\Service\Registrar\RegistrarInterface;
 
 final class FileProcessor
 {
     public function __construct(
-        private CommentDetector $commentDetector,
-        private CommentExtractor $commentExtractor,
-        private RegistrarInterface $registrar,
+        private CommentRegistrar $commentRegistrar,
         private Saver $saver,
         private Tokenizer $tokenizer,
     ) {
@@ -24,37 +19,10 @@ final class FileProcessor
     public function process(\SplFileInfo $file): void
     {
         $tokens = $this->tokenizer->tokenize($file);
-        $commentTokens = $this->commentDetector->filter($tokens);
-        if (!$commentTokens) {
-            return;
-        }
-
-        if (!$this->registerTodos($commentTokens)) {
+        if (!$this->commentRegistrar->register($tokens)) {
             return;
         }
 
         $this->saver->save($file, $tokens);
-    }
-
-    /**
-     * @param \PhpToken[] $tokens
-     */
-    private function registerTodos(array $tokens): bool
-    {
-        $hasNewTodo = false;
-        foreach ($tokens as $token) {
-            $commentParts = $this->commentExtractor->extract($token->text);
-            foreach ($commentParts->getTodos() as $commentPart) {
-                if ($this->registrar->isRegistered($commentPart)) {
-                    continue;
-                }
-                $this->registrar->register($commentPart);
-                $hasNewTodo = true;
-            }
-
-            $token->text = $commentParts->getContent();
-        }
-
-        return $hasNewTodo;
     }
 }
